backend/requirements.txt
django==5.1.*
djangorestframework==3.15.*
psycopg2-binary==2.9.*
django-cors-headers==4.*
django-filter==24.*
redis==5.*
Pillow==11.*
gunicorn==23.*
